+++ RELEASE_NOTES.md
@@ -1,3 +1,6 @@
+### 0.7.4 (2015-07-21)
+* [[WAZ.Storage][Improvment] Log storage request information from StorageException in TableBatchOperationAdapter.
+
 ### 0.7.3 (2015-07-08)
 * [LanguageExtensions][BugFix] EmptyMemoryStream.Get() returns reusable (non disposable) stream instance.